Remove legacy code and refactor kit, ability, and passive systems
