+++ plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/SuperSmashMobsBrawl.kt
@@ -6,9 +6,8 @@ import dev.betrix.superSmashMobsBrawl.commands.KitCommand
 import dev.betrix.superSmashMobsBrawl.commands.LeaveCommand
 import dev.betrix.superSmashMobsBrawl.commands.QueueCommand
 import dev.betrix.superSmashMobsBrawl.commands.resolvers.MinigameDefinitionArgument
-import dev.betrix.superSmashMobsBrawl.extensions.hasPassive
+import dev.betrix.superSmashMobsBrawl.extensions.hasPassiveId
 import dev.betrix.superSmashMobsBrawl.models.brawlData.MinigameDef
-import dev.betrix.superSmashMobsBrawl.passives.definitions.HungerPassiveDefinition
 import dev.betrix.superSmashMobsBrawl.services.DataService
 import dev.betrix.superSmashMobsBrawl.services.DebugService
 import dev.betrix.superSmashMobsBrawl.services.HotbarService
@@ -99,7 +98,7 @@ class SuperSmashMobsBrawl : SuspendingJavaPlugin() {
         event<FoodLevelChangeEvent> {
             if (entity is Player) {
                 val player = entity as Player
-                if (!player.hasPassive(HungerPassiveDefinition)) {
+                if (!player.hasPassiveId("hunger")) {
                     player.feed()
                     isCancelled = true
                 }
