--- plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/abilities/AbilitySpec.kt
@@ -1,9 +1,0 @@
-package dev.betrix.superSmashMobsBrawl.abilities
-
-/** Runtime specification for an ability, built from data files. */
-data class AbilitySpec(
-    val id: String,
-    val name: String,
-    val metadata: AbilityMetadata,
-    val displayItem: String? = null,
-)
