+++ plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/extensions/player.kt
@@ -1,6 +1,5 @@
 package dev.betrix.superSmashMobsBrawl.extensions
 
-import dev.betrix.superSmashMobsBrawl.passives.definitions.PassiveDefinition
 import dev.betrix.superSmashMobsBrawl.services.DebugService
 import dev.betrix.superSmashMobsBrawl.services.KitService
 import dev.betrix.superSmashMobsBrawl.utils.mm
@@ -28,14 +27,13 @@ fun Player.sendDebugMessage(message: String) {
     sendMessage(mm("<gray>[DEBUG] $message</gray>"))
 }
 
-fun Player.hasPassive(definition: PassiveDefinition): Boolean {
+fun Player.hasPassiveId(passiveId: String): Boolean {
     // New data-driven kits
     val brawlKit = KitService.getBrawlKit(this)
     if (brawlKit != null) {
-        return brawlKit.passives.any { it.id == definition.id }
+        return brawlKit.passives.any { it.id == passiveId }
     }
 
-    // Legacy kits
-    val kit = KitService.getKitInstance(this)
-    return kit?.passiveInstances?.find { it.definition.id == definition.id } != null
+    // Legacy kits (none now)
+    return false
 }
