--- plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/kits/definitions/KitDefinition.kt
@@ -1,13 +1,0 @@
-package dev.betrix.superSmashMobsBrawl.kits.definitions
-
-import dev.betrix.superSmashMobsBrawl.kits.KitMetadata
-import dev.betrix.superSmashMobsBrawl.kits.instances.KitInstance
-import org.bukkit.entity.Player
-
-abstract class KitDefinition {
-    abstract val name: String
-    abstract val id: String
-    abstract val metadata: KitMetadata
-
-    abstract fun createInstance(player: Player): KitInstance
-}
