--- plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/kits/instances/CreeperKitInstance.kt
@@ -1,13 +1,0 @@
-package dev.betrix.superSmashMobsBrawl.kits.instances
-
-import dev.betrix.superSmashMobsBrawl.disguises.CreeperDisguise
-import dev.betrix.superSmashMobsBrawl.kits.definitions.KitDefinition
-import org.bukkit.entity.Player
-
-class CreeperKitInstance(definition: KitDefinition, player: Player) :
-    KitInstance(definition, player) {
-    override fun setup() {
-        disguise = CreeperDisguise(player)
-        super.setup()
-    }
-}
