--- plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/passives/PassiveSpec.kt
@@ -1,4 +1,0 @@
-package dev.betrix.superSmashMobsBrawl.passives
-
-/** Runtime specification for a passive, built from data files. */
-data class PassiveSpec(val id: String, val name: String, val metadata: PassiveMetadata)
