--- plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/passives/instances/PassiveInstance.kt
@@ -1,22 +1,0 @@
-package dev.betrix.superSmashMobsBrawl.passives.instances
-
-import dev.betrix.superSmashMobsBrawl.Manageable
-import dev.betrix.superSmashMobsBrawl.extensions.sendDebugMessage
-import dev.betrix.superSmashMobsBrawl.passives.definitions.PassiveDefinition
-import org.bukkit.entity.Player
-
-abstract class PassiveInstance(val definition: PassiveDefinition, val player: Player) :
-    Manageable() {
-
-    override fun setup() {
-        super.setup()
-
-        player.sendDebugMessage("You have been given the ${definition.name} passive")
-    }
-
-    override fun teardown() {
-        super.teardown()
-
-        player.sendDebugMessage("The ${definition.name} passive has been removed")
-    }
-}
