--- plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/registries/KitRegistry.kt
@@ -1,22 +1,0 @@
-package dev.betrix.superSmashMobsBrawl.registries
-
-import dev.betrix.superSmashMobsBrawl.kits.definitions.CreeperKitDefinition
-import dev.betrix.superSmashMobsBrawl.kits.definitions.KitDefinition
-import dev.betrix.superSmashMobsBrawl.kits.definitions.SkeletonKitDefinition
-
-object KitRegistry {
-    private val definitions = mutableMapOf<String, KitDefinition>()
-
-    fun register(definition: KitDefinition) {
-        definitions[definition.id] = definition
-    }
-
-    fun getDefinition(id: String): KitDefinition? = definitions[id]
-
-    fun getAllDefinitions(): Collection<KitDefinition> = definitions.values
-
-    init {
-        register(CreeperKitDefinition)
-        register(SkeletonKitDefinition)
-    }
-}
