+++ plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/services/HotbarService.kt
@@ -1,10 +1,8 @@
 package dev.betrix.superSmashMobsBrawl.services
 
 import dev.betrix.superSmashMobsBrawl.abilities.AbilityUsageType
-import dev.betrix.superSmashMobsBrawl.abilities.instances.AbilityInstance
 import dev.betrix.superSmashMobsBrawl.brawl.BrawlAbility
 import dev.betrix.superSmashMobsBrawl.brawl.BrawlKit
-import dev.betrix.superSmashMobsBrawl.kits.instances.KitInstance
 import gg.flyte.twilight.scheduler.repeatingTask
 import org.bukkit.NamespacedKey
 import org.bukkit.entity.Player
@@ -21,89 +19,29 @@ object HotbarService : Listener {
     private lateinit var plugin: JavaPlugin
     private lateinit var abilityKey: NamespacedKey
 
-    // Track which players have kits with hotbar items
     private val playersWithKits = mutableSetOf<Player>()
 
     fun initialize(plugin: JavaPlugin) {
         this.plugin = plugin
         this.abilityKey = NamespacedKey(plugin, "ability_id")
         plugin.server.pluginManager.registerEvents(this, plugin)
-
-        // Start periodic hotbar updates (every 1 second)
         repeatingTask(20) { updateAllPlayerHotbars() }
     }
 
-    /** Sets up hotbar items for a kit instance */
-    fun setupHotbarItems(kitInstance: KitInstance) {
-        val player = kitInstance.player
-        playersWithKits.add(player)
-
-        // Clear hotbar first
-        clearHotbar(player)
-
-        // Add ability items to hotbar
-        kitInstance.abilityInstances.forEach { abilityInstance ->
-            addAbilityToHotbar(player, abilityInstance)
-        }
-
-        player.updateInventory()
-    }
-
-    /** Sets up hotbar items for a brawl kit */
     fun setupHotbarItems(kit: BrawlKit) {
         val player = kit.player
         playersWithKits.add(player)
-
         clearHotbar(player)
-
         kit.abilities.forEach { ability -> addAbilityToHotbar(player, ability) }
-
         player.updateInventory()
     }
 
-    /** Clears hotbar items for a player when their kit is removed */
     fun clearHotbarItems(player: Player) {
         playersWithKits.remove(player)
         clearHotbar(player)
         player.updateInventory()
     }
 
-    /** Adds an ability item to the player's hotbar (legacy AbilityInstance) */
-    private fun addAbilityToHotbar(player: Player, abilityInstance: AbilityInstance) {
-        val metadata = abilityInstance.definition.metadata
-        val slot = metadata.hotbarItemSlot
-
-        // Clone the item to avoid modifying the original
-        val item = metadata.hotbarItem.clone()
-
-        // Add persistent data to identify which ability this item represents
-        val meta = item.itemMeta
-        if (meta != null) {
-            meta.persistentDataContainer.set(
-                abilityKey,
-                PersistentDataType.STRING,
-                abilityInstance.definition.id,
-            )
-
-            // Add cooldown info to lore if on cooldown
-            val lore = mutableListOf<String>()
-            lore.add("§7${metadata.description}")
-            if (abilityInstance.isOnCooldown()) {
-                lore.add("§cCooldown: ${abilityInstance.getRemainingCooldown()}s")
-            } else {
-                lore.add("§aReady to use!")
-            }
-            lore.add("§eRight-click to activate")
-
-            meta.lore = lore
-            item.itemMeta = meta
-        }
-
-        // Place item in the specified slot
-        player.inventory.setItem(slot, item)
-    }
-
-    /** Adds an ability item to the player's hotbar (new BrawlAbility) */
     private fun addAbilityToHotbar(player: Player, ability: BrawlAbility) {
         val metadata = ability.metadata
         val slot = metadata.hotbarItemSlot
@@ -112,7 +50,6 @@ object HotbarService : Listener {
         val meta = item.itemMeta
         if (meta != null) {
             meta.persistentDataContainer.set(abilityKey, PersistentDataType.STRING, ability.id)
-
             val lore = mutableListOf<String>()
             lore.add("§7${metadata.description}")
             if (ability.isOnCooldown()) {
@@ -128,154 +65,70 @@ object HotbarService : Listener {
         player.inventory.setItem(slot, item)
     }
 
-    /** Clears the player's hotbar (slots 0-8) */
     private fun clearHotbar(player: Player) {
         for (i in 0..8) {
             player.inventory.setItem(i, null)
         }
     }
 
-    /** Updates hotbar items to reflect current cooldown states */
-    fun updateHotbarItems(kitInstance: KitInstance) {
-        val player = kitInstance.player
-        if (!playersWithKits.contains(player)) return
-
-        kitInstance.abilityInstances.forEach { abilityInstance ->
-            addAbilityToHotbar(player, abilityInstance)
-        }
-
-        player.updateInventory()
-    }
-
-    /** Updates hotbar items to reflect current cooldown states for BrawlKit */
     fun updateHotbarItems(kit: BrawlKit) {
         val player = kit.player
         if (!playersWithKits.contains(player)) return
-
         kit.abilities.forEach { ability -> addAbilityToHotbar(player, ability) }
-
         player.updateInventory()
     }
 
-    /** Gets the ability ID from an item's persistent data */
     private fun getAbilityId(item: ItemStack?): String? {
         if (item == null || !item.hasItemMeta()) return null
-
         val meta = item.itemMeta ?: return null
         return meta.persistentDataContainer.get(abilityKey, PersistentDataType.STRING)
-    }
-
-    /** Finds a kit instance for a player (legacy) */
-    private fun getLegacyKitInstance(player: Player): KitInstance? {
-        return KitService.getKitInstance(player)
-    }
-
-    /** Finds a brawl kit for a player */
-    private fun getBrawlKit(player: Player): BrawlKit? {
-        return KitService.getBrawlKit(player)
     }
 
     @EventHandler
     fun onPlayerInteract(event: PlayerInteractEvent) {
         val player = event.player
-
-        // Only handle players with kits
-        if (!playersWithKits.contains(player)) {
-            return
-        }
-
+        if (!playersWithKits.contains(player)) return
         val item = event.item ?: return
         val abilityId = getAbilityId(item) ?: return
 
-        // Prefer new data-driven kit
-        val brawlKit = getBrawlKit(player)
-        if (brawlKit != null) {
-            val ability = brawlKit.abilities.find { it.id == abilityId } ?: return
+        val kit = KitService.getBrawlKit(player) ?: return
+        val ability = kit.abilities.find { it.id == abilityId } ?: return
 
-            when (ability.metadata.usageType) {
-                AbilityUsageType.LEFT_CLICK -> {
-                    if (
-                        event.action != Action.LEFT_CLICK_AIR &&
-                            event.action != Action.LEFT_CLICK_BLOCK
-                    ) {
-                        return
-                    }
-                }
-                AbilityUsageType.RIGHT_CLICK -> {
-                    if (
-                        event.action != Action.RIGHT_CLICK_AIR &&
-                            event.action != Action.RIGHT_CLICK_BLOCK
-                    ) {
-                        return
-                    }
-                }
-            }
-
-            event.isCancelled = true
-
-            if (ability.canActivate()) {
-                ability.activate()
-                updateHotbarItems(brawlKit)
-            }
-
-            return
-        }
-
-        // Fallback to legacy
-        val kitInstance = getLegacyKitInstance(player) ?: return
-        val abilityInstance =
-            kitInstance.abilityInstances.find { it.definition.id == abilityId } ?: return
-
-        when (abilityInstance.definition.metadata.usageType) {
+        when (ability.metadata.usageType) {
             AbilityUsageType.LEFT_CLICK -> {
                 if (
                     event.action != Action.LEFT_CLICK_AIR && event.action != Action.LEFT_CLICK_BLOCK
-                ) {
+                )
                     return
-                }
             }
             AbilityUsageType.RIGHT_CLICK -> {
                 if (
                     event.action != Action.RIGHT_CLICK_AIR &&
                         event.action != Action.RIGHT_CLICK_BLOCK
-                ) {
+                )
                     return
-                }
             }
         }
 
-        // Cancel the event to prevent normal item behavior
         event.isCancelled = true
 
-        // Attempt to activate the ability
-        if (abilityInstance.canActivate()) {
-            abilityInstance.activate()
-
-            // Update hotbar items to reflect new cooldown state
-            updateHotbarItems(kitInstance)
+        if (ability.canActivate()) {
+            ability.activate()
+            updateHotbarItems(kit)
         }
     }
 
     @EventHandler
     fun onPlayerQuit(event: PlayerQuitEvent) {
-        // Clean up tracking when player disconnects
         playersWithKits.remove(event.player)
     }
 
-    /** Updates hotbar items for all players with kits */
     private fun updateAllPlayerHotbars() {
         playersWithKits.toList().forEach { player ->
-            val kit = getBrawlKit(player)
+            val kit = KitService.getBrawlKit(player)
             if (kit != null) {
                 updateHotbarItems(kit)
-                return@forEach
-            }
-
-            val legacy = getLegacyKitInstance(player)
-            if (legacy != null) {
-                updateHotbarItems(legacy)
             } else {
-                // Player no longer has a kit, remove from tracking
                 playersWithKits.remove(player)
             }
         }
